+++ database/migrations/2024_09_01_113047_create_credit_repayment_account_mandates_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('credit_customer_debit_mandates', function (Blueprint $table) {
-            $table->id();
+        if (! Schema::hasTable('credit_customer_debit_mandates')) {
+            Schema::create('credit_customer_debit_mandates', function (Blueprint $table) {
+                $table->id();
 
-            $table->string('identifier')->nullable()->unique(); // system generated e.g ACCMNDT-VENDOR_CODE-YEARMONTHDAY-PRIMARY_ID i.e ACCMNDT-10MG-20240901-23
-            $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade'); // BusinessType::VENDOR
-            $table->foreignId('customer_id')->constrained('credit_customers')->onDelete('cascade');
+                $table->string('identifier')->nullable()->unique(); // system generated e.g ACCMNDT-VENDOR_CODE-YEARMONTHDAY-PRIMARY_ID i.e ACCMNDT-10MG-20240901-23
+                $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade'); // BusinessType::VENDOR
+                $table->foreignId('customer_id')->constrained('credit_customers')->onDelete('cascade');
 
-            // ...mandate information
-            // TODO: Check paystack documentation to run a new migration that add additional columns
+                // ...mandate information
+                // TODO: Check paystack documentation to run a new migration that add additional columns
 
-            $table->index(['business_id', 'customer_id', 'identifier'], 'mandate_business_customer_identifier_unique');
-            $table->timestamps();
-        });
+                $table->index(['business_id', 'customer_id', 'identifier'], 'mandate_business_customer_identifier_unique');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
